Move learn status related logic to a custom useLearnStatus hook
